Working PoC using Cloudflare credentials and DNS challenge.
